Refactor repository manifest conditional handling
As noted in #15 there were some edge cases around the `vcsrepo` task
which could cause the provisioner to mutate the remotes in existing
wordpress-develop checkouts. This could be confusing, but it was not
obvious how to address the issue using Puppet resource relationships.

This PR uses Puppet facts to gather information about the environment
before running the manifest, so that we can conditionally skip any
unwanted or unneeded tasks.